Parallel Commands, High Speed Turn, Current Limit
Changed auto modes to have parallel functions for arms drop and elevator
height.  Added a parallel command for scale auto to move to switch
height in parallel, reducing time for max height.  Upped current
limiting on drive train,  currently at 35A for peak and 25A for
flatline.  Created high speed turn function, but it is probably
unnecessary because we can use regular turn function, but with different
P ?I? D values.  Need to determine if cube acquisition can be trusted to
Ded reckoning, or if Vision is needed.